git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@4600 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpe2/src/logging/include/clog_mem.h
@@ -12,7 +12,7 @@
 #if defined(MPIR_MEMDEBUG)
 /* Enable memory tracing.  This requires MPICH's mpid/util/tr2.c codes */
 #include "mpimem.h"             /* Chameleon memory debugging stuff */
-#define MALLOC(a)       MPID_trmalloc((size_t)(a),__LINE__,__FILE__)
+#define MALLOC(a)       MPID_trmalloc((unsigned)(a),__LINE__,__FILE__)
 #define FREE(a)         MPID_trfree(a,__LINE__,__FILE__)
 #define REALLOC(a,b)    realloc(a,b)
 #else
