+++ src/mpe2/src/wrappers/src/mpe_proff.c
@@ -75,7 +75,7 @@ char *alloca ();
 */
 #if defined(MPIR_MEMDEBUG)
 /* Enable memory tracing.  This requires MPICH's mpid/util/tr2.c codes */
-#define MALLOC(a)    MPID_trmalloc((size_t)(a),__LINE__,__FILE__)
+#define MALLOC(a)    MPID_trmalloc((unsigned)(a),__LINE__,__FILE__)
 #define FREE(a)      MPID_trfree(a,__LINE__,__FILE__)
 
 #else
